refactor: near structions
